feat(imposto-renda): expandir extração com 17 novos campos DEDUÇÕES e IMPOSTO PAGO
- Adicionar 10 campos individuais de DEDUÇÕES ao IncomeTaxInfo DTO:
  Contrib. Prev. Oficial, RRA, Complementar, Dependentes, Instrução,
  Médicas, Pensão Judicial, Escritura, Pensão RRA, Livro Caixa

- Adicionar 7 campos individuais de IMPOSTO PAGO ao IncomeTaxInfo DTO:
  Retido Fonte Dependentes, Carnê-Leão Titular/Dependentes,
  Imposto Complementar, Exterior, Lei 11.033, Retido RRA

- Adicionar padrões regex e lógica de extração para todos os novos campos
- Adicionar persistência PayrollEntry com códigos IR_DEDUCOES_* e IR_IMPOSTO_*
- Corrigir regex tolerante a encoding para Ano-Calendário e Exercício
- Corrigir GlobalExceptionHandler para respostas já commitadas (erro 500)
- Refatorar uso de BigDecimal nos DTOs de consolidação

+++ src/main/java/br/com/verticelabs/pdfprocessor/interfaces/consolidation/dto/ConsolidatedResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -19,7 +20,6 @@ public class ConsolidatedResponse {
     private Set<String> anos; // ["2016", "2017", "2018"]
     private List<String> meses; // ["01", "02", ..., "12"]
     private List<ConsolidationRow> rubricas;
-    private Map<String, Double> totaisMensais; // formato: "2017-01" -> 700.31
-    private Double totalGeral;
+    private Map<String, BigDecimal> totaisMensais; // formato: "2017-01" -> 700.31
+    private BigDecimal totalGeral;
 }
-
