+++ src/main/java/br/com/verticelabs/pdfprocessor/interfaces/consolidation/dto/ConsolidationRow.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
@@ -14,7 +15,6 @@ import java.util.Map;
 public class ConsolidationRow {
     private String codigo;
     private String descricao;
-    private Map<String, Double> valores; // formato: "2017-01" -> 424.10
-    private Double total;
+    private Map<String, BigDecimal> valores; // formato: "2017-01" -> 424.10
+    private BigDecimal total;
 }
-
