+++ src/main/java/br/com/verticelabs/pdfprocessor/interfaces/entries/dto/EntryResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -15,12 +17,11 @@ public class EntryResponse {
     private String rubricaCodigo;
     private String rubricaDescricao;
     private String referencia; // YYYY-MM
-    private Double valor;
+    private BigDecimal valor;
     private String origem; // CAIXA, FUNCEF ou INCOME_TAX
     private Integer pagina;
-    
+
     // Campos opcionais da rubrica (quando disponível)
     private String rubricaCategoria; // Categoria da rubrica (ex: "Administrativa", "Extraordinária")
     private Boolean rubricaAtivo; // Indica se a rubrica está ativa
 }
-
